# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VoidScan Security Operations Centre API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /file/hash:
    post:
      operationId: fileHash
      tags: [file]
      summary: SHA256/MD5 Generator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileDataInput"
      responses:
        "200":
          description: File hash results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HashResult"
  /file/hash-compare:
    post:
      operationId: fileHashCompare
      tags: [file]
      summary: Hash Comparison
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HashCompareInput"
      responses:
        "200":
          description: Hash comparison result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResult"
  /file/metadata:
    post:
      operationId: fileMetadata
      tags: [file]
      summary: File Metadata Extraction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileDataInput"
      responses:
        "200":
          description: File metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileMetadataResult"
  /file/entropy:
    post:
      operationId: fileEntropy
      tags: [file]
      summary: Entropy Analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileDataInput"
      responses:
        "200":
          description: Entropy analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntropyResult"
  /file/malware-hash:
    post:
      operationId: fileMalwareHash
      tags: [file]
      summary: Malware Hash Lookup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetInput"
      responses:
        "200":
          description: Malware hash lookup result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MalwareHashResult"
  /file/yara:
    post:
      operationId: fileYara
      tags: [file]
      summary: YARA Rule Scanner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/YaraInput"
      responses:
        "200":
          description: YARA scan result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YaraResult"
  /file/signature:
    post:
      operationId: fileSignature
      tags: [file]
      summary: File Signature Verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileDataInput"
      responses:
        "200":
          description: File signature result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignatureResult"

  # ─── Password Security ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    TargetInput:
      type: object
      required: [target]
      properties:
        target:
          type: string
    FileDataInput:
      type: object
      required: [filename, data]
      properties:
        filename:
          type: string
        data:
          type: string
          description: Base64 encoded file content
        mimeType:
          type: ["string", "null"]
...
    HashCompareInput:
      type: object
      required: [hash1, hash2]
      properties:
        hash1:
          type: string
        hash2:
          type: string
    YaraInput:
      type: object
      required: [data, rule]
      properties:
        data:
          type: string
          description: Base64 encoded content to scan
        rule:
          type: string
          description: YARA rule text
    GenericResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        data:
          type: object
          additionalProperties: true
        error:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
...
    HashResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        filename:
          type: ["string", "null"]
        md5:
          type: ["string", "null"]
        sha1:
          type: ["string", "null"]
        sha256:
          type: ["string", "null"]
        sha512:
          type: ["string", "null"]
        size:
          type: ["integer", "null"]
        error:
          type: ["string", "null"]
    FileMetadataResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        filename:
          type: ["string", "null"]
        size:
          type: ["integer", "null"]
        mimeType:
          type: ["string", "null"]
        extension:
          type: ["string", "null"]
        magicBytes:
          type: ["string", "null"]
        encoding:
          type: ["string", "null"]
        error:
          type: ["string", "null"]
    EntropyResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        filename:
          type: ["string", "null"]
        entropy:
          type: ["number", "null"]
        interpretation:
          type: ["string", "null"]
        isSuspicious:
          type: ["boolean", "null"]
        byteDistribution:
          type: object
          additionalProperties: true
        error:
          type: ["string", "null"]
    MalwareHashResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        hash:
          type: ["string", "null"]
        found:
          type: ["boolean", "null"]
        malwareName:
          type: ["string", "null"]
        malwareType:
          type: ["string", "null"]
        sources:
          type: array
          items:
            type: string
        error:
          type: ["string", "null"]
    YaraResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        matched:
          type: ["boolean", "null"]
        matchedRules:
          type: array
          items:
            type: string
        error:
          type: ["string", "null"]
    SignatureResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        filename:
          type: ["string", "null"]
        detectedType:
          type: ["string", "null"]
        magicBytes:
          type: ["string", "null"]
        isValid:
          type: ["boolean", "null"]
        details:
          type: ["string", "null"]
        error:
          type: ["string", "null"]
...