# lib/api-zod/src/generated/types/entropyResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VoidScan Security Operations Centre API
 * OpenAPI spec version: 0.1.0
 */
import type { EntropyResultByteDistribution } from './entropyResultByteDistribution';

export interface EntropyResult {
  success: boolean;
  /** @nullable */
  filename?: string | null;
  /** @nullable */
  entropy?: number | null;
  /** @nullable */
  interpretation?: string | null;
  /** @nullable */
  isSuspicious?: boolean | null;
  byteDistribution?: EntropyResultByteDistribution;
  /** @nullable */
  error?: string | null;
}