# lib/api-zod/src/generated/types/genericResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VoidScan Security Operations Centre API
 * OpenAPI spec version: 0.1.0
 */
import type { GenericResultData } from './genericResultData';

export interface GenericResult {
  success: boolean;
  data?: GenericResultData;
  /** @nullable */
  error?: string | null;
  /** @nullable */
  message?: string | null;
}