# lib/api-zod/src/generated/types/yaraResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VoidScan Security Operations Centre API
 * OpenAPI spec version: 0.1.0
 */

export interface YaraResult {
  success: boolean;
  /** @nullable */
  matched?: boolean | null;
  matchedRules?: string[];
  /** @nullable */
  error?: string | null;
}